+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 - Now requires Go 1.18 to build.
 - Directives in the config file are now case-insensitive.
+- If a route has multiple upstreams an error is now raised, instead of
+  silently ignoring some of them.